Subscriptions service now working.
Fetching membership creates any component of the subscription framework if it is missing.
User can now manage their membership through a billing portal by clicking manage subscription.
Membership will stay active after cancelling until the end of the period paid for.
Creates a new checkout session if there are no subscriptions or if any of the subscriptions are inactive/expired.